assets/graph/requirements.txt
kafka-python 
matplotlib 
plotly
nbformat